fix: drop a game pose when the world no longer holds its entity
From the review on #256.

A hot reload clears the registry and reads the scene again, and EnTT then
hands the same entity numbers out with a new version. StepMotion held the
old handles, so the next step wrote a stale pose. Either the entity was
gone, which World::set_local does not check, or a new entity had taken its
number and quietly got somebody else's transform.

The runtime clears the poses at the reload now, beside the selection it
already clears for the same reason. StepMotion also drops an entity the
world no longer holds, so a caller that forgets cannot corrupt a scene.

Simulated seconds reach the game as a double as well. Narrowing at that
boundary threw away what the double was for. The claim that a float stops
counting steps "after a few hours" was wrong: it resolves 1/60 until about
three days of running.

The docs that came with the branch had sentences over 25 words, which the
writing style forbids.

+++ sandbox/game.h
@@ -139,13 +139,15 @@ namespace sandbox {
      *
      * @param world The world to run.
      * @param seconds Simulated seconds since the game started, not since the
-     *                last step.
+     *                last step. A double, because the runtime counts steps in
+     *                one and narrowing here would throw that away at the
+     *                boundary.
      * @param motion Where to record what moved, so a frame between two steps
      *               draws a blend rather than the newest step. Every entity
      *               this moves is recorded in it.
      * @return How many entities moved.
      */
-    std::size_t update(engine::scene::World& world, float seconds,
+    std::size_t update(engine::scene::World& world, double seconds,
                        engine::scene::StepMotion& motion);
 
 } // namespace sandbox
